+++ PontoAll.WebAPI/Services/Entities/UserService.cs
@@ -6,6 +6,7 @@ using PontoAll.WebAPI.Objects.Models;
 using PontoAll.WebAPI.Controllers;
 using PontoAll.WebAPI.Services.Interfaces;
 using static PontoAll.WebAPI.Controllers.UserController;
+using PontoAll.WebAPI.Services.Utils;
 
 namespace PontoAll.WebAPI.Services.Entities;
 
@@ -38,13 +39,13 @@ public class UserService : GenericService<User, UserDTO>, IUserService
 
     public async Task Create(UserDTO dto)
     {
-        UserValidator.Validate(dto);
+        CpfCnpjValidator.Validate(dto);
         await base.Create(dto);
     }
 
     public async Task Update(UserDTO dto, int id)
     {
-        UserValidator.Validate(dto);
+        CpfCnpjValidator.Validate(dto);
         await base.Update(dto, id);
     }
 }
